Restructure the Application class
This is a first step to simplifying how `Application` instances are
created and used, with the ultimate goal being to simplify how they're
tested. Rather than instantiating the application with a handful of
settings, it's now instantiated with a settings object. This will allow
for easier settings overrides, plus it serves to ensure that the
settings are available everywhere.

This change brings with some changes that will affect any services using
the ingestion.service library. The signature of the callback function
exposed through `process.run` needs to change to accept an application
instance as its first parameter.

Additionally each application instance now receives a name. This will
typically be the name of the service, but could eventually be utilized
to uniquely identify each running process.

This change also brings along with it some restructuring of how the
Kafka client is handled. There are no major technical changes being
made, only structural ones affecting the layout of the code.