Switch to using Kafka plugin
In order to allow the plugin system to set default values for settings,
the way settings work inside `Application` needs to change. In addition
to switching all Kafka-related functionality over to the new plugin,
settings are now converted from an object -- typically they'll be a
module, but classes work, too -- to a dict.