Add an application registry
There will be times when an application instance is needed but not
directly available. By maintaining a registry of applications, they can
be fetched at any time.

The most recent application can be set and retrieved through the
registry instance's `current_application` property.

TODO: The registry probably needs to be cleaned up when an application
falls out of scope.