Rename to Verona
It's almost as if this is a brand new "initial commit." There are some
significant changes being made here. Because of this, the version number
is being reset to 0.1.0. This is possible because...

The library is being renamed from ingestion.service to Verona. This was
prompted by a desire to use the library for services outside the
Ingestion Pipeline. Typing the name back to Ingestion just didn't make
sense.

In addition to the rename, `Application` is being restructured a little
to allow for more external control. The class's signature is changing a
bit to accept a consumer (more on this later) and an error callback. The
`settings` argument is being made optional. Settings can now be set,
extended, and overridden after the instance has been created. Part of
the change for this is that settings are being handled through a new
`Config` class that allows for extension from other a `dict` or an
object in addition to direct assignment.

Next, the package namespace is being dropped. As nice as it was, it was
just begging for name collisions with different plugins using modules
with the same name.

The next change is that the hard requirement on Kafka is being removed.
Any object that implements the consumer interface will work. This will
make it easier to test different services.

Lastly, and perhaps most importantly, some documentation is being added.