Add a class for creating lazy proxies
A new class, `ProxyObject`, is being added; it acts as a proxy around
any object or callable. The main advantage of the class is that it will
delay the execution of the callable until it's attributes are accessed.
It will allow for references to be created at runtime to things that
don't yet exist.

When using a callable it does not cache the result.