Change Consumer Interface to be anything iterable
When Henson consumes through its consumer instance, it uses the instance
as an iterator, with the exception that the consumer must provide a
method called `read` which returns the iterator. The Consumer Interface
can be simplified to iterate over the consumer directly. This means that
custom classes and won't be needed just to wrap around iterators.