What rolls down stairs...
While there is still work to be done, this is the beginning of logging
in Henson. Applications will now provided a logger that can be used to
send all messages through the same logger.

Some initial log messages are being added to `run_forever`. They will
track the startup and shutdown of the application (as info messages) as
well as any exceptions that occur trying to read a message. This will
probably be the extent of the messaging in `Application` with the
possible exception of debugging messages added to track the full startup
process.