Remove current_application
With recent changes to existing Henson extensions, current_application
is not a useful construct within Henson anymore. Extensions that provide
additional functionality to Henson applications now require explicit
references to their applications in order to operate properly.
current_application and supporting code are removed but may return in
the future if a compelling use case is found.