Add docs-requirements.txt
Henson's docs require shinxcontrib-napoleon to build. ReadTheDocs is
able to build documentation using napoleon, but only if a requirements
file is provided. A new requirements file is added to allow ReadTheDocs
to build this documentation, and individual requirements specified in
tox.ini are hoisted and replaced with `-rdocs-requirements.txt` to avoid
duplication.