Add application autodetection to run command
Typically, a service built with Henson will have a single instance of an
application defined in a module. When using the run command, this
instance can be automatically chosen, removing the need to pass the
attribute name to `henson run`.

While adding this feature, some of `henson run`'s errors were made more
clear. Import and attribute errors are no longer hidden by the CLI and
bubble up.