Add pre- and postprocessing callbacks
Applications may wish to use some sort of middleware to manipulate
incoming messages before they are passed into the application and the
results of processing them. To achieve this two new parameters are being
added to `Application`: `message_preprocessors` and
`result_postprocessors`. Each of these parameters takes a list of
callables that will be executed at the appropriate time.

Also, with the addition of the new callbacks, a new Callbacks page is
being added to the documentation to give a brief overview of what each
one does.