Centralize version information
Version information is moved from the two different places it's
currently stored (setup.py and docs/conf.py) into a `henson.__version__`
attribute.

PEP 396 [0] recommends that version information for packages should be
declared in an attribute called `__version__` and that the version
attribute used in setup.py should be derived from `__version__`. In
addition, to ensure that version information is consistent everywhere
it's used.

[0] https://www.python.org/dev/peps/pep-0396/#specification