Henson-Retry
Retry is a plugin for Henson that will allow an application to
automatically retry a message that failed to process. It does this by
registering an error callback that runs before all others. The callback,
however, will only run when the exception is of a supported type and the
retry attempt threshold and timeout haven't been reached.

This is the first of what could be many contrib packages for Henson.

As part of this change, the control flow of `Application.run_forever` is
being modified to support multiple error callbacks. These callbacks will
be called when a message fails to process, rather than when Henson fails
to receive a message from the consumer.