Process messages asynchronously
We've decided to make Henson process messages asynchronously. Since 3.4
is the lowest support version of Python, asynchronous support is being
added through `asyncio`.

When Henson runs an application is will create one future for the
consumer. The consumer needs to expose a coroutine named `read`
(**note** this is a backwards incompatible change). The coroutine should
return the next message to be processed. (Internally, the message will
be placed on an asynchronous queue that will be read by the processing
futures.)

`Application.run_forever` is gaining a new argument named `num_workers`.
This represents the number of futures that Henson will start to process
messages read in through the consumer. The `callback` and any callables
provided for `error_callbacks`, `message_preprocessors`, and
`result_postprocessors` must be coroutines.

These changes also simplify the flow used inside `run_forever` and allow
for much cleaner error handling.

To round things off, the documentation is being updated to account for
the new consumer interface.

This should be considered a first pass and not a stable implementation.