Add exception to stop processing a message
Henson now provides an `Abort` exception to allow preprocessors to tell
the application to stop processing a message. The exception can also be
used by postprocessors to stop processing a result. Any time an `Abort`
is caught, it will be logged along with information about who raised it.

Along with the `Abort`, a couple of new test fixtures are being
introduced: `queue` and `cancelled_future`. `queue` will provide an
`asyncio.Queue` instance to a test. `cancelled_future` will provide a
future that has already been cancelled. This can be useful for tests
that want to test processing until the consumer has been cancelled.

Closes #50