Reorganize test fixtures
With the number of fixtures declared in `conftests.py`, we've decided
that module-specific fixtures should be defined directly in the module.
If the fixture ever has a need outside the module, it can be moved.

The fixtures for the CLI tests are being moved into the CLI test module.
Some of the tests require the source of `MockApplication`. An instance
of the class is provided through the `test_app` fixture; its source can
be obtained from the instance's type.

The `callback` fixture is mostly leftover from prior to the introduction
of `asyncio` and is currently only used by the Henson-Retry tests. It's
being moved to the retry module until it is removed entirely.

References #58