Register callbacks, now with startup and teardown
Henson would benefit from callbacks that can be run when an application
starts up and tears itself down (e.g., opening and closing connections).
As more types of callbacks are added , the signature for `Application`'s
initializer grows harder to read. Rather than continuing to add new
callbacks this way, all callbacks will now be registered through
instance methods. These methods can also be used as decorators.

The callback registration methods are:

- `application_startup`
- `application_teardown`
- `message_preprocessor`
- `result_postprocessor`
- `error_callback`

Additionally, some changes are being made to the control flow of
`run_forever` to accommodate its expanded functionality. The two most
notable changes are the explicit passing of a `loop` argument to pretty
much anything that accepts it and a change to the signature of
`_consume` to accept a future that it can cancel when the consumer
raises an `Abort` exception.

Fixes #53