Acknowledge messages
Depending on the source of incoming messages, acknowledgement of the
message may be needed to prevent other consumers (or even the same one)
from receiving the same message. A `message_acknowledgement` callback is
being added to allow Henson to handle this.

Without this change, consumers would have to handle acknowledgement
themselves. This would be problematic if an application failed in the
middle of processing a message the consumer had already acknowledged.

Closes #52