Rename callback registers
Three of the callback registers are being renamed. `application_startup`
will be known as `startup`, `application_teardown` will be known as
`teardown`, and `error_callback` will be known as `error`. This makes
their names less redundant and easier to read.

The callback documentation is being updated to not only reflect these
new names, but callbacks that have been added and other changes
introduced as part of Henson 0.5.

References #69

Closes #73