Update Henson-Retry to work with Henson 0.5
Henson-Retry is being updated to use coroutines. (This will make
implementing the delay easier since a non-blocking sleep can be used.)
It's also being updated to raise `henson.exceptions.Abort` rather than
`StopIteration`. This requires an update to `henson.base.Application`
and introduces a new test. This test caught an error in how exceptions
are logged.

The way callbacks are registered has changed, so that's being updated as
well. This will need to be updated when the callback decorator names are
finalized.

References #73

Closes #58