Add debug mode
Debugging with asyncio can be tricky. Without enabling its debug
mode--this is generally done by setting the ``PYTHONASYNCIODEBUG``
environment variable--exceptions inside tasks can be swallowed.

To make it easier to see this information, a debug mode is being added
to Henson. It can be enabled through the command line, the call to
`run_forever`, and the `DEBUG` application configuration setting.

Closes #85