Fix reloader behavior
Currently, the reloader fails to start the application because the
thread spawned for it has no event loop. To fix that issue, a new event
loop is created and passed to ``app.run_forever``. Additionally, this
loop is explicitly passed to all functions that take a ``loop`` argument
to ensure that the correct loop is always being used. Finally,
``run_forever`` now calls ``set_event_loop`` to ensure that any other
async code running in the reloader thread (e.g. Henson extensions) that
may rely on the default event loop are not left without a loop to use.