Use an include for the example in the docs
Sphinx has a `literalinclude` directive that can be used to pull in code
from a file. A file is being added to house the `FileConsumer` example.
It will be included by both `index.rst` and `interface.rst` so that the
code doesn't need to be kept up-to-date in two places.