Enable extending the command line interface
Henson's command line interface is simple, offering just a `run`
command. This is a great starting point, but there is a world of
potential for extending the command line through extensions.
Unfortunately doing that requires knowing the implementation details of
the CLI and some trickery to register new commands.

A function, `register_commands`, is being added to make it easier for
extensions to register new commands. In addition to the new function, a
new argument, `--app` is being added to the CLI. Specifying an
application will result in it being imported, allowing any of its
extensions to register new commands. Commands can choose to receive the
application through an argument named `app`.

The usage for a new command is generated from the function's signature.
This is done much in the same way as argh does, but the `@arg` decorator
isn't needed for functions accepting `app`. `register_commands` will
handle this automatically.

The other deviation from argh's behavior is that namespaces are required
to extend Henson's CLI.