Change style of CLI help text
While docstrings are great for people reading the usage of a function,
especially when combined with Sphinx, they don't always translate well
to the command line.

    $ henson run --help

properly displays the docstring, although all of the details appear
associated with the command, not with its arguments.

    $ henson --help

doesn't even properly display the docstring (i.e., all newlines are
removed).

To display the help text in a way that better matches what argh
provides, the documentation style is being changed. The docstring itself
is being reduced to just a description of what the command will do.
Arguments are being documented through function annotations.