Add verbosity controls to the run command
A mutually exclusive set of arguments is being added to the `henson run`
command: `--quiet` and `--verbose`. Providing one of the arguments will
alter the level of logging used by Henson's logger. While `run` only
supports one level of quiet, both arguments are counts. The more times
`--verbose` is specified, the lower the log level threshold will be.

As part of this, `run`'s logging calls are being raised from `DEBUG` to
`INFO` so that they will be logged upon the use of a single `--verbose`.

Closes #131