Add a Sphinx plugin to autogenerate CLI docs
There is a Sphinx extension named autoprogram that will autogenerate
Sphinx documentation for a CLI. This works great for Henson itself.
Unfortunately it falls short when documenting Henson extensions because
the extension needs to register its subcommands in order for them to be
associated with the CLI.

A new contrib plugin is being added (`henson.contrib.sphinx`). It wraps
around autoprogram and handles importing the extension automatically.