Add support for uvloop
uvloop is an alternate event loop for use with asyncio. While it's easy
to provide a specific event loop for a Henson application to use when
calling `run_forever` directly, there's no easy way to do so when using
the `henson run` CLI. Because uvloop is a drop-in replacement, it's easy
to add support for it directly into Henson by checking for it at run
time.

I had originally implemented this by setting the event loop policy at
compile time. While that approach feels a tad bit cleaner (easier to
remove if we need to stop support uvloop), this approach is much easier
to test.