Set the event loop when running with reloader
Normally Henson applications can take care of setting their own event
loops. When running with the reloader, however, a loop is needed to
ensure that the same event loop is always used each time the application
is started. The `_new_event_loop` function added in 58cac31 can be used
for this.