Rename an incorrectly named argument
The `Application._process` method accepts an `asyncio.futures.Future` as
its first argument. The future signatures and docstring, however,
referred to it as an `asyncio.tasks.Task`. This was left over from the
original asynchronous implementation introduced in 13f510f. The `Task`
created there was changed to a `Future` in c55cd32. This change should
have been made then.