Provide a default start command for CLI extensions
When building the documentation for Henson plugins, sphinx-autoprogram
picks up `sphinx-build` as the program. This requires setting the `prog`
option whenever a plugin uses the Henson Sphinx extension to document
its CLI. This change will provide a default value of
`henson --app APP_PATH` so that plugin authors can omit it.