Halt processing after an exception
Currently the only way to stop the application (without the use of
`SIGTERM` or `SIGKILL`) is for the consumer to shut down (either
naturally or through an exception). When a `_process` future fails,
`future` will be stopped. Because the event loop is only monitoring the
`consume` future, however, the application will keep running with the
consumer blocked by a full queue.

This normally isn't an issue as most of `_process` happens inside a
try/except block. There are three instances, however, where exceptions
will not be handled:

    - result postprocessing: This happens inside the try's else block.
    - error callbacks: This happens inside the try's except block.
      `Abort` will be caught, but nothing else.
    - message acknowledgement: This happens inside the try's finally
      block.

By changing the call to `run_until_complete` to monitor a future that
includes both the consumer and processing futures, an exception raised
by either will end the application.