Reduce the number of futures used for consuming
`run_forever` creates a future to keep track of whether or not the
consumer is still running. It then creates a separate task from the
future. If the application is stopped by some means other than the
first future completing on its own, the task often raises a
`RuntimeError` will trying to do something after the event loop has been
closed.

There is no need to keep track of both futures. A task can be made from
the `_consume` coroutine. When `_consume` exits, the task will be marked
done, when `consumer.cancel` is called the task will be stopped cleanly.