Use a Travis CI build matrix
In order to use both Python 3.5 and 3.6 on Travis CI, both need to be
included in the list of Python versions. This would result in the entire
build matrix being run using both versions. 3.5 jobs would fail on the
3.6 image and 3.6 jobs would fail on the 3.5 image.

While more verbose than what we had before, this build matrix allows us
to run each job with the correct version of Python.